last fixing
Took 2 hours 35 minutes

+++ app/Http/Controllers/Profile/PersonalDataController.php
@@ -30,6 +30,11 @@ class PersonalDataController extends Controller
     public function destroy(Request $request)
     {
         $user = auth()->user();
+
+        if ($user->hasRole('Admin')) {
+            return response()->json(['message' => 'Вы не можете удалить пользователя с правами администратора.'], 403);
+        }
+
         $request->session()->invalidate();
         $user->delete();
